in which we get a (kinda) working proof-of-concept
== DETAILS
So, the biggest issue was that.. memory in the default heap is apparently
not executable, so if you try to jump to it you get a freeze. Thanks to
@Maschell for pointing that out.

Once I got that straightened out, the next issue is that my hello_world()
function was returning an empty string. Well, I think what's *actually*
happening is that the library isn't getting fully initialized, i.e. the code
that's supposed to set all the default values for things isn't being run,
so the memory address where "Hello, world!" is supposed to live is zeroed
out, which translates to an empty string.

For now, I modified the function to return an int, and updated the caller
appropriately. And that works! \o/