Do some more fleshing out
I tried to invoke the entrypoint as specified in the elf header,
but this ended up crashing. The rpl spec seems to be.. doing nothing?

Anyway, I modified it to look up the `rpl_entry` symbol and invoke it
with `OS_DYNLOAD_LOADED` on init and `OS_DYNLOAD_UNLOADED` during
teardown.

I modified the library very slightly to add a little complexity and see
if everything still works (it does!).

I suspect where things will get dicey is when the rpl needs to do things
that involve libc functions.